forwarding: extract tunnel outer-resolution + neighbor lookup into tunnel.rs
Pure code-motion split of forwarding/mod.rs (#5650). Move the tunnel
outer-header forwarding resolution (resolve_tunnel_outer,
resolve_tunnel_forwarding_resolution, outer_neighbor_ifindex) and the
neighbor-entry lookup/parse helpers (lookup_neighbor_entry,
parse_neighbor_entries) into forwarding/tunnel.rs.

Functions are moved verbatim; no logic, signature, or hot-path change.
pub(super) becomes pub(in crate::afxdp) so the mod.rs glob re-export
keeps all call sites resolving.

Validation: cargo build clean (0 errors).
